Use hypothesis to create example brief response data
Creates a dictionary strategy to create brief response data example.
Sets default settings to run each example only once, so that our test
times are pretty much unchanged, but it's always possible to change
the value to get all the benefits of generative testing.

(I wasn't really planning on committing this, but it actually caught
a bug in the whitespace filter, so I'm keeping it.)