Validate the number of answers in the brief response
Checks that the number of answers matches the number of questions
in the brief for both essential and nice to have requirements.

This assumes that requirement answers are either boolean or have a
similar uniform structure (eg "Yes"/"No" enum values) and each item
is validated by the JSON schema, so the only invalid answer type not
caught by schema validation is the wrong number of answers.