Add list brief responses endpoint
Can be filtered by supplier_id and brief_id.

Returns a paginated list only if not filtered, otherwise returns
all responses at once (similar to what we do with supplier services
to avoid having to deal with pagination in frontend code).

Returns a 200 with no results for non-integer supplier_id and brief_id.
Different endpoints handle this differently (some 400, some will raise
a DataError). We don't really expect those requests from the frontend
apps, so I'm opting for the simplest way to handle them - which is
ignoring the invalid value.