Add utils.get_int_or_400 helper function
Can be used to check optional request.args parameters that have to
be integer if present (e.g. supplier_id, brief_id, user_id).

Updates list briefs to return 400 if user_id is not integer.
Updates list brief responses to return 400 if brief_id or supplier_id
are not integer.