Add gunicorn to requirements.txt for PaaS deploys
Add gunicorn to requirements.txt for running the app on Cloud Foundry
(possibly not ideal see [1]).

Explicitly add cffi to requirements so that Cloud Foundry spots it and
adds libffi.

There doesnt' seem to be an easy way to use a different requirements
file for Cloud Foundry, so we'll have to add the dependencies
everywhere.

[1] https://github.com/etianen/django-herokuapp/issues/9