Don't drop None values when stripping whitespace
Changes utils function to only remove empty strings from lists.
None can be either a valid or invalid value depending on the field,
so the validation logic should be encoded in the field schema. If
None is valid, the schema should define it as one of the possible
item values, otherwise a validation error should be returned.