Move clarification question validation to schemas
Move the brief clarification question validation of question and answer
into JSON schemas.

The previous approach made it difficult to return multiple errors at the
same time. Moving the validation out of field validators into a
`validate` method as is used on `BriefResponse` makes it easier to
return multiple errors. Standardising the error messages so that they
match those from other endpoints is an effort to make the API more
internally consistent.