Add ability to filter briefs by status
Uses the sqlalchemy hybrid_property with a separate function to be
used in query expressions.

This means we need to reimplement the logic twice (both for getting
the status property for python and for filtering based on status in
the database), but the `.status` property behaves like a regular
column for the rest of the API code.

http://docs.sqlalchemy.org/en/latest/orm/extensions/hybrid.html#module-sqlalchemy.ext.hybrid