Add Brief.clarificaiton_quesitons_are_closed property
Checking whether the close date has passed seems like it will be
the most common operation related to close dates. When checking
whether applications are closed the app can check the brief status
instead, but for clarification questions the frontend apps would
need to parse the date first.

This adds a helper property to Brief objects and serialized data
that returns true if clarification close date is in the future and
false otherwise.