Allow service IDs to be generated for DOS
Remove the restriction on serivce IDs that only allows them to be
generated for g-cloud-7.

This generates a string of integers 15 digits long.
The length has been reduced to 15 digits to avoid Excel translating the
number to exponential format.
I've changed from exponential notation to raising to the power because
the exponential notation creates floats and the larger float overflows.