Don't allow updating or deleting closed briefs
Changes brief status checks to look for non-'draft' briefs instead
of checking for 'live' status only. This way, 'live' and 'closed'
briefs aren't allowed to be updated or removed and the check will
cover any similar statuses added in the future.

Fixes the missing `test_` in test name for live brief delete.