Change brief application and questions closing times to 23:59
Setting the closing date and time to 23:59 + 14 days means that we
can show the date part of the timestamp separately since it's the
last day applications/questions are open as opposed to the first
day they're closed for 00:00 dates. Making the change in the API
allows frontend apps to display the date without any modifications.