Migration to add `data` to `lots` table with info about units
This is a JSON blob that will store additional data about the lot that is needed by the frontend applications but that we will never need to index or search by (so it doesn't need a new column for each item).

This also adds the "singular" and "plural" units for each lot - i.e. service/supplier/lab.