Abort migrations check if a version is present more than once
list_migrations checks whether we have more than one branch in the
list of migration versions. Since we've switched to a new revision
naming convention, pull requests open at the same time are likely
to use the same revision number when adding new migrations (ie if
the current latest migration is '500' two pull requests adding a
new one are both likely to use '510' for their revision id). When
this happens, alembic excludes on of the migrations from the tree
and doesn't consider it to be a multiple branches issue. Instead,
it prints out a warning. In order to make sure the tests fail when
this happens we transform the warning into an exception inside the
list_migrations script. This is similar to running the python
interpreter with `-W error`.