Change method for extracting domain from email
The previous method would cause an error if there was more than one @
symbol in the email address. The new method allows many @ symbols to be
used in an email address whilst still grabbing just the domain.

A valid email only has one @ symbol, however making this allowance in
the method means that an invalid email address using more than one @
symbol can pass this check, allowing it to fail gracefully when checked
against the pattern defined on the users schema.