add, use DMMutableContainer (derived from the sqlalchemy.ext.mutable extension) in attempt to make @validates more reliable
when @validates is used on a JSON field it doesn't pick up mutations of the orm-mapped objects for re-validation. this implementation
of mutable containers does a little better, advertizing edits of the top-level container to the validator, but is still not perfect
as it doesn't catch edits to deep-nested mutable objects - that might be a further exercise.