Return uploader info in agreementDetails if UserId
One of the fields in `agreementDetails` is an `uploaderUserId`,
which refers to a user but doesn't provide our frontend apps with
very much useful information.
To resolve this, we've updated the `.serialize()` method on a
SupplierFramework to always return the name and email address
of the uploader user if the uploaderUserId field exists in the
agreementDetails and the user id exists in our database.