Validate different required fields based on progress
Firstly, we don't want to validate or update `agreementDetails`
unless the framework in question has an
agreement_details_version.

If it does, then we want smarter validation logic.

Due to `agreementDetails` being built up over several pages,
we want to require that progressively more fields exist in
`agreementDetails` without losing any of the fields that existed
previously.

This means that we want to pass in larger and larger lists of
`required` fields depending on how far the user has made it
through the flow of signing and returning their framework
agreement.

More concretely:
- Step 1 requires a `singerName` and a `signerRole`
- Step 2 requires an `uploaderUserId` *in addition* to the signer
details fields (which should have already been added in Step 1)