Make validation logic stricter and add tests
Recently, the `framework_agreement_version` (string) field
attached to frameworks was changed to a JSON field called
`framework_agreement_details`, containing a
`frameworkAgreementVersion` key.

Rewrote the existing logic to accomodate for this, and then added
a check that if `agreementReturned=True` for a framework with
`framework_agreement_details`, a framework version must exist
as well.

Finally, dropped `framework_agreement_details` from being
validated because it doesn't exist in the JSON schema.