Allow setting frameworkAgreementVersion
It is technically possible that returning an agreement could set
an incorrect `frameworkAgreementVersion` in the `agreementDetails`
field.
If this happens (which it shouldn't), we want to be able to
manually correct the frameworkAgreementVersion.

Also, this allows us to add `frameworkAgreementVersion` to the
`agreementDetails` JSON schema, which means that it now describes
every field that `agreementDetails` can contain.