Unify all of the `agreementDetails` logic
Previously, we were validating `agreementDetails` near the top of
the view (after the other validations), and then updating
the actual data near the bottom of the view (after the other
updates).

However, there are a lot of checks to do when validating or
setting `agreementDetails` for a given SupplierFramework, and
it ended up as two separate, scary-looking blocks with a lot
of duplication for no real gain.

By unifying the validation and assignment logic, we can at least
limit the blast radius of our terrible new code.