Fail updating agreementDetails with nonexistent user
Previous behaviour was to save any `uploaderUserId` whether it
corresponded to a real user or not.

Users are certain to exist at the point when the `uploaderUserId`
is successfully set for the first time, but it is conceivable that
the user could be deleted after some time, leaving us with an
`agreementDetails` field for which most updates would fail (eg,
trying to set countersigner details).

We don't want to be adding bad information to our database, but
we also don't want other, unrelated updates to fail myteriously
(which they might if we were validating the user id every time).