Commit supplier data to the database before indexing
When created for the first time Supplier objects contain attribute
values as they were assigned, so if `supplier.code` is assigned a
string value it's not transformed into integer before the instance
is stored. Which means that document is indexed before database
constraints are checked and values like `"005"` are transformed to
integer `5`. When the same document is updated however, code read
from the database is an integer value, which doesn't match Elasticsearch
`_id = "005"`, so a new document is created when indexing the update.

Moving `db.commit` before Elasticsearch indexing makes sure that the
document in Elasticsearch index matches the value stored in the
database.