Add logic in model to withdraw brief
If 'withdrawn' is passed through as a status, set the
`withdrawn_at` field to the current time.
If `withdrawn_at` is set, return a status of 'withdrawn'
Once a brief has been withdrawn, it is no longer
possible to modify the status.
Add `withdrawnAt` to Brief.serialize

Also added some tests around this to the model.