Replace brief status route with publish route
We wish to restrict status changes so we can explicitly define
what updates are allowed to a brief's status. We introduce the
"publish" route to change a brief from 'draft' to 'live'.

You are no longer able to pass in a 'status' in the payload.

We refactored our tests and removed a test that checks invalid
statuses are not allowed because this is no longer applicable.