Remove functionality to unpublish a brief
This functionality to unpublish a brief was previously added in a
 commit (a8526d80f43819fb48fdbbebba876ed2598ae3e0). We wish to
now remove this functionality. This is because we will no longer
allow briefs to be unpublished (i.e. to go from 'live' to 'draft')
and instead they must be withdrawn ('live' to 'withdrawn').

We intend to later on add a route to create a new draft from a
withdrawn brief.