Add back in older update status route for backwards compatibility
In previous commits we have removed the /briefs/<brief_id>/status
route which would allow the user to update the status to their
choice. This was removed as we decided to take a stricter approach
to letting users update the status. However in retrospect, so we
can deploy this API change without breaking our other apps we will
need to allow support from this older functionality until we can
replace it with the newer functionality in our apps.