Only index `g-cloud` services
We've been indexing all edited services as long as:
(a) they are live (published)
(b) the framework they belong to is live

Unfortunately, this meant that DOS services were getting into
our g-cloud index.

New logic also checks:
(c) the framework they belong to is a `g-cloud` framework

We talked about several solutions, including one where we only
index services if their `Framework.framework` value matches an
alias in our elasticsearch indices.  This *would* work and it
would save us from hardcoding `g-cloud` in our application code,
but we don't think the scale of the problem really calls for
this level of abstraction.

Currently, we're only using elasticsearch for one set of services
and all of those services are `g-cloud` services.  Done.