Trying to write tests for the index method, which previously had no tests.
First problem I ran into was that the tests are testing
an actual `Service` object, so they I couldn’t just use
the existing DOS/G-Cloud fixtures.

So I went for the method to create dummy services.

Then, I needed a G-Cloud 8 framework, so I created a class
that inherits from `BaseApplicationTest` but adds a G-Cloud 8
framework in the setup and removes it in the teardown.

Unfortunately, my tests were *still* failing because the
initial statuses of the frameworks were all wrong.
G5 and G6 are the only live frameworks; G7, DOS, and G8 are
in standstill, pending, and live, respectively.

So this is where I completely lost the plot.
I expanded the test class in this really clunky way to:
- override all of the initial framework statuses
- revert them all to their initial statuses in the teardown

If we start changing stuff about frameworks between
tests, then other tests start to fail.

Basically, this is the fundamental problem.
Our tests are all sharing the same set of `Framework` database
objects, which we used to set and update in migrations but we
don't any longer.  Now that we've stopped, the default state of
all of our frameworks no longer reflect the world as it
presently exists.

It’s also pretty inconvenient to create a totally new service
object.  Before we have a `Service`, we need:
- a `Supplier`
- a `Framework`
- a `Lot` that belongs to that `Framework`

And there’s no straightforward way to create new `Framework`
or `Lot` objects and then get rid of them at the end of the test.

This pull request sucks and I don’t think we should merge it.
But we should figure out what to do about these tests.