Refactor brief copying functionality into model and generalise
- Changed 'create-draft-from-withdrawn-brief' route to be 'copy'
- Moved copy functionality into the Brief model
- Removed restriction that brief must be withdrawn to copy
- Removed functionality that a copied brief would have it's title
changed to be 'Copy of [original_brief_name]' as decided by Sarah.
This was because we would have an edge case if the brief title was
too close to the field length limit and we would not be able to
prepend 'Copy of ' without breaking that, or would have to offer
differing and unclear behaviour to the user by only prepending the
string if the brief title would still be a valid length.
- Change `update_brief_status_by_action` route to use POST instead
of PUT
- Small refactoring to Brief model to improve logic