First attempt to refactor framework fixtures
Pulled out the getting/setting logic of our framework fixture into
a regular function and then set up a couple of fixtures which pass
in some variable parameters to hardcoded methods.
This works and it's easy to see how to add more fixtures as
needed, but this way involves lots of hardcoding stuff into the
file with the fixture definitions, as it doesn't seem possible to
generate fixtures on the fly or very cleanly pass variables
through to our fixtures (at least, not in the py.test examples).

[Passing parameters documented here](https://pytest.org/dev/fixture.html#parametrizing-a-fixture)