Return a function from a fixture
This is kind of a hack that lets us define an inner
function that accepts parameters inside of our fixture.
This way, we can directly call the fixure in our test and pass in
a couple of parameters.
It's totally weird because you're not meant to be calling fixtures
directly inside the test, so it's taking advantage of
incidental behaviour.

[Based on this stackoverflow answer](http://stackoverflow.com/questions/18011902/py-test-pass-a-parameter-to-a-fixture-function).