Create general fixture for framework updates
It turns out it's possible to pass parameters to fixtures
using a decorator (thus overriding any default parameters).

As such, we can generalise our `live_framework` fixture to a more
generic `update_framework` fixture which, by default, works
as a drop-in replacement.
Thereafter, when we need an expired framework, we can just
use the `mark.parameterize()` decorator to override the status
parameter above the test that wants it.

This way, we don't have to create new fixtures per scenario,
but can just tweak what's important in our general fixture.

The idea comes from [this stackoverflow answer](http://stackoverflow.com/questions/28027363/any-way-to-pass-parameters-into-pytest-fixture).