Add index_service tests
Create a new service in the test and assign it a framework in the
state we need it to be using our new framework fixtures.

It feels a bit odd how some test parameters are mocks being patched
in while others are updating database framework values, but it
did seem easy enough to put them together this way.