Break out SupplierFrameworkResponse test
In our `SupplierFrameworkUpdates` class, we had one test that
was verifying all of the fields we expected to find in a
`SupplierFramework` were there.
This meant it was creating a `SupplierFramework` object in the
setup for all other tests and storing it in the database.

As this test diverged quite a lot from the other tests in the
class, I broke it into its own class with one test method in it.
Removed the database insertion from the setup step because we
only want to do it once, and removed the `JSONUpdateTestMixin`
inheritance because we're not posting any data.

Note that there's no teardown method needed because our inherited
`teardown_database` method will wipe the database of anything
that isn't a `Framework` or a `Lot`.