Named fixures per framework per status
Biting the bullet and creating hardcoded frameworks fixtures
which either add a new framework in a particular status or
update the status of an already-existing framework as needed.

Passing in parameters from the tests themselves seemed
like more overhead than it was worth, although it would be
reaasonably straightforward to create a generic `add_framework()`
fixture that added arbitrary new frameworks with passed-in
slugs.

Sort of ironic this was esssentially @allait's solution to begin with.