SupplierFramework: add migration adding agreed_variations json field
added as a field separate from agreement_details because the latter is exposed and near-arbitrarily
updateable (proviso passing the json schema validation) through update_supplier_framework_details.
seeing as we're going to the trouble of implementing a special endpoint for agreeing to variations
that can do so in a transactionally-safe way, it's probably better not to let clients circumvent all
that safety through manual tinkering.

beyond the race argument, it was decided that as we wanted a specific audit event type for variation
agreements, a specific endpoint for agreement made more sense than sniffing for what the caller is
up to in update_supplier_framework_details to determine an event type. and if we're providing a
bespoke endpoint, it's probably best not to allow it to be manually tinkered with through other means,
which it would be if it lived in agreement_details.