Make a list of variation keys explicitly
Python 2 returns dict keys as a list, but in Python 3 it returns `dict_keys` which is a view and not a list.  This results in a `dict_keys() is not JSON serializable` error when running under Python 3.

By explicitly making a list here it works in both Python 2 and Python 3.