[#128752313] Create more generic framework update route
Modify existing update route which would only accept payloads
containing both `status` and `clarificationQuestionsOpen`.
New route is more generic (keys can be passed together or
seperately), but only allows whitelisted keys through.

At the moment, our whitelisted keys are:
- `status`
- `clarificationQuestionsOpen`
- `frameworkAgreementDetails`

Since `frameworkAgreementDetails` is a JSON field, we want to
validate the JSON that gets passed through against a schema if
it exists, and then overwrite the existing details.
This means potentially losing data if you don't include all of
the existing keys (since it's not an update), but the other
option -- just updating the existing dictionary, would mean
that we can never delete keys.
We don't really use this route too often, so it seems like it
should be okay.