Only validate clarification_questions_open if status exists
It turns out that validating `clarification_questions_open` based
on the value of `self.status` was resulting in a weird race
condition in python3 where, for example, creating a new framework
would sometimes throw the validation error.  If the clarification
questions key was attempted to be set before a status existed at
all, the validation would fail and no framework would be created.

In lieu of a way to enforce that one field is always validated
before the other (certainly
[this gist doesn't do this](https://gist.github.com/matrixise/6417293)),
I'm only validating the `clarification_questions_open` field
*if* the `status` field currently has a value.

Note that it's _possible_ to pass in both a `"status": "live"` and
a `"clarificationQuestionsOpen": True` in the JSON update and, if
the framework's status is originally set to `"open"`, the change
might go through.
Seems like a super low risk thing, though, and easy to reverse.