Create more generic framework fixture
We have been trying to move away from using Frameworks that
look exactly like ones that we actually use, and so an example
framework is a step in this direction.

Note that `Framework.framework` has to be either `g-cloud` or
`dos`, so we currently have to use one of those values.