Add fixture parameterization to the agreement tests
It's possible to parameterize fixtures that aren't used directly by
the test, so when making fixtures that depend on one another we can
still set parameters on the base ones.

We're using this to set additional fields on the framework fixture
for framework agreement tests.

This adds a helper `@fixture_params` decorator to provide an easier
interface for the common case.