Add sign framework agreement route
Route to handle signing of frameworks agreements (regardless of if
they have a `frameworkAgreementVersion` or not).

We try and copy the existing behaviour for signing a framework
agreement through the supplier framework routes (such as ability
to resign a framework agreement). However we have not copied
across the ability to unsign a framework agreement that has already
been signed (as it's not currently being used in our apps and this
behaviour would shortly be removed anyway as we continue to
refactor and have multiple `frameworkAgreement`s.