Use `frameworkSlug` instead of `frameworkId` for `FrameworkAgreements`
Previously we were sending back a `frameworkId` in our
`FrameworkAgreement.serialize` method. We change this to make it
consistent with the `SupplierFramework` (that does not make the
framework ID accessible but instead uses the framework slug for
identifying a framework).

To faciliate this change, have adapted the `supplier_framework`
fixture to return a more flexible result.