Use fixtures for database set up for test_suppliers
We change to use fixtures. We also remove some tests that are
checking setting a `SupplierFramework` as countersigned. This is
because this functionality is not yet being used in any of our apps
and will in the future be done instead in the `FrameworkAgreement`
model. Removing the actual functionality to set a framework
agreement as countersigned through the supplier views will be
removed in a later commit to match.