Save framework agreement information in `framework_agreement` table
We still set framework agreement data in the supplier views but now
we will save the data to the `framework_agreement` table.

If a record already exists in the `framework_agreement` table then
we will update it. If a record does not yet exist in the
`framework_agreement` table, then we will create a new record using
any existing data from the `supplier_frameworks` table.

This enforces a one to one relationship between a `SupplierFramework`
and a `FrameworkAgreement`.

This will later allow us to migrate all our data from the
`supplier_frameworks` table to the `framework_agreements` table.