Fix get framework suppliers by agreement returned
Now that framework agreement information can be stored in two
different places (`SupplierFramework` or `FrameworkAgreement`) we
need to change the `get_framework_suppliers` route to look in the
correct place for framework agreement information. It should use
the information from `FrameworkAgreements` if available rather than
`SupplierFramework`. This also uses the current assumption that
`FrameworkAgreement` will have a one to one relationship to
`SupplierFramework`.