Fix user export bug which references framework agreements
To export data on a user for the `export_users_for_framework` route,
we include if the framework agreement had been returned. This
would use just the `SupplierFramework.agreement_returned_at` property
to decide. Now that we will be having framework agreement information
stored in the `framework_agreement` table, this will break this
functionality. Instead of duplicating the logic of which data source
to look at (`SupplierFramework` or `FrameworkAgreement`), we can
look at the serialized `SupplierFramework` which already has done
this logic for us.