Migrate framework agreements to `framework_agreements`
For supplier_framework rows that contain details of returning a framework agreement (indicated
by non null `agreement_details` or `agreement_returned_at`) and do not yet have a
corresponding row in the `framework_agreements` table.

Note, we are moving framework agreements regardless of if they have a framework
agreement version or not.