Keep SupplierFrameworks consistent with FrameworkAgreements
While we are using a combination of both tables to store agreement details we should keep the data in both places consistent.

This will fix a bug with unsetting `agreementReturnedAt` where it gets updated in FrameworkAgreement but not in SupplierFrameworks, and so continues to be returned by the list framework agreements endpoint for `agreement_returned=true` even though it has been unset.