Test finding correct framework agreements
This is related a a bug fixed in commit 29349255a491f8d7066dd449b88e336e1c09bce2.

Previously, the code would look for any `FrameworkAgreement`
belonging to a supplier and update it.  This meant that if

- a framework agreement existed for a supplier
- a framework agreement update came in for a _different_ framework

the framework agreement for the first (incorrect) framework would
_always_ be returned and updated.
A new framework agreement would never be created.

This test verifies that even if a framework agreement exists for
another framework, a new framework agreement is always created.