Add `status` hybrid property to `FrameworkAgreement` and `serialize`
Although there is anticipated to be a two step process to countersigning
a framework agreement (first setting `countersigned_agreement_returned_at`
and then later setting `countersigned_agreement_path`), we have
decided that if the `countersigned_agreement_returned_at` is set
then the status of `FrameworkAgreement` will be `countersigned`. If
frontend apps need the file path to exist for them to reveal that a
framework agreement has been countersigned then they will need to
explicitely check this themselves rather than just relying on the
status.