[#130991743] Remove application_status from framework users export
Current implementation of application_status check requires N+1 queries
to request draft services for each framework supplier. This is timing
out due to the number of suppliers in production.

As a temporary workaround, since we're mainly interested in exports
for live frameworks at the moment we're removing application status
values from the export.