Add temporary countersign route needed for one off script
Temporary route needed so we can set a user defined timestamp for an agreement
to be countersigned (the real route that is still to be created will not allow
the user to define the timestamp but will take the current time). This route
will also allow us to set the countersigned path. After a one off refactoring script
is run then this route will no longer be needed and we can delete it.