Add `SupplierFramework.current_framework_agreement` property
Allows looking up the current framework agreement for the supplier.
Right now the only way to access this is either using `agreements[0]`
from the view (which would break in the future) or calling
`supplier_framework.serialize()`.

Adding a property encapsulates the current agreement logic in a model
property, which should make it easy to use in other views.