Simplify users export query and avoid calling sf.serialize
Export endpoint is still timing out in production. Looking at the
recent changes it seems that serializing supplier_framework records
to check for framework agreement status might be slowing down the
API response.

The query time in production is ~6s, so it doesn't feel like it's
the main cause of 30s timeouts.

This simplifies the database query and replaces `sf.serialize` with
an attribute check.