Add endpoint to countersign agreement
Currently we're not entirely sure where the name of the countersignee is
coming from. Most likely the framework table. This end point will check
to see if there is a name in the framework_agreement_details blob, and
store it if there is. It also adds the name of the person who pushed the
button so it records something either way.

This can be fixed up when we know where the countersignee is coming
from.