Return updaters name and email in serialized supplier_framework
The supplier_framework needs to supply more information about who
clicked the button in the admin app to countersign the agreement.

This info is to be used in the admin app front end.

The users ID is stored in the framework_agreement
countersigned_agreement_details column. The serialize method on the
supplier_framework looks at this ID in the framework_agreement and
returns that users details.