Add 'approved' as a framework_agreement status
If an admin has approved a framework agreement (which timestamps it
countersigned_agreement_returned_at field) but the script to generate
it's signed document (which adds the path to the framework agreement)
hasn't been run yet, its status needs to be 'approved' instead of
countersigned. It's status is only upgraded to 'countersigned' once that
path exists.