`SupplierFramework.current_framework_agreement` picks relevant FA
`SupplierFramework.current_framework_agreement` now supports
multiple framework agreements per supplier framework.

For the moment, we include drafts in the SupplierFramework to keep our interface the same whilst we refactor
the frontend to allow multiple framework agreements per supplier framework. This means that if a draft is created
after any other framework agreement then it must take precident (as the supplier frontend for the moment only
knows about one framework agreement (the current framework agreement) we must expose the draft so it can be
edited as they complete the signing flow).

Note, as we have no timestamp for framework agreements being created we instead have to use the highest `id`.
To support this we want to restrict the order in which our framework agreements are returned by ordering the
backref by `FrameworkAgreement.id` instead of the default SQLAlchemy order of most recently updated.

Also note, this means that for a short while, if someone has
signed/countersigned a supplier framework and then starts a new draft then the supplier framework will return to
'draft' status and will lose knowledge of the previous signing.

After the supplier frontend has been refactored so that the signing flow deals with a framework agreement rather
than a suqpplier framework then we will be able to ignore drafts and not include them in the current framework
 agreement