Bugfix for DOS single-service lots
The functionality that prevents suppliers from creating multiple services on "single-service" lots was not checking the framework that the service belonged to.

So suppliers with a DOS service in a single-servcie lot were prevented from creating a DOS2 service on the same lot.

This fixes it to be conditional on lot too, so DOS suppliers will be able to submit their services for DOS2.