Validate only requested questions when creating brief response
We change from `enforce_required` being true to false. We instead
validate just the questions which are requested. Note, this fixed
a bug in the current validation logic where even if fields were
not supposed to be checked then we would still do so, e.g.  we would
always run the validation logic to check the number of essentials
provided equalled the number of essentials attributed to the brief.