Bugfix (temporary) - ordering framework agreements
The SQLAlchemy `.order_by` clause was not behaving as expected - rather than ordering on the already-joined framework_agreements table it was doing a second outer join with no relationships specified and ordering on that, which gives undefined results.  (Sometimes it works and sometimes it doesn't).

This TEMPORARILY replaces the broken `order-by` by ordering the SupplierFrameworks using Python, in advance of fixing the API to exclude drafts from `current_framework_agreement`, at which point we can properly sort this using SQL once an expression for `current_framework_agreement` can be more easily defined.