Build error messages for dynamic_list questions
Errors for dynamic_list questions need to be linked to the field
that is failing the validation: either the primary question or a
follow-up question (depending on the value of the primary question).
eg if we answer 'yes' to the question about having the essential
requirement, we then have to have a value for 'evidence'.

To achieve this, we're putting one extra key into the `errors_map`
dictionary indicating which field failed the validation.