Validate brief responses with correct schema
We have the following three cases:
- feature flag is off
- brief is created before the feature flag
- brief is created after the feature flag

For the first two cases, we should be validating our data against
the legacy brief response schema. For the later case we should be
validating against the new brief response schema.

Note, it's not just good enough to validate against the correct
JSON schema. We also need to only do further complex checks in the
`validate` method, such as checking the number of `essentialRequirements`
for the legacy schema.