Add tests for oneOf validator error translation
For most validation tests we're using example services with actual
schemas to check the errors.

Since there's no schemas with dynamic list or followup questions yet
we need to use a different approach for testing oneOf validator errors.

So instead of running get_validation_error, we split the translation
logic into a separate function and test that using manually made
ValidationError instances.

This is a much more granular test since it only tests one part of
the translation without touching anything else, but it's not tied
to a schema so if it changes the tests need to be updated too.