Write tests for creating a brief response for both feature flag cases
We originally had one set of tests for creating a brief response.
However now we have three cases to consider:
- feature flag is off
- brief is created before the feature flag
- brief is created after the feature flag

For the first two cases, we should be validating our data against
the legacy brief response schema. For the later case we should be
validating against the new brief response schema.

Note, we have not included any tests for the case when the feature
flag is off. Is this OK?